requirements.txt
django==5.0.6
djangorestframework==3.15.2
markdown==3.3.6